Move deployment snapshot yaml handling to models/save_and_load.py
