+++ src/deploy_tools/snapshot.py
@@ -1,10 +1,8 @@
 import logging
 
-import yaml
-
 from .layout import Layout
 from .models.deployment import Deployment, DeploymentSettings
-from .models.load import load_from_yaml
+from .models.save_and_load import load_from_yaml, save_as_yaml
 
 logger = logging.getLogger(__name__)
 
@@ -20,9 +18,9 @@ def create_snapshot(deployment: Deployment, layout: Layout) -> None:
     configuration when a compare, validate or sync process is run.
     """
     _backup_snapshot(layout)
+
     logger.debug("Creating snapshot: %s", layout.deployment_snapshot_path)
-    with open(layout.deployment_snapshot_path, "w") as f:
-        yaml.safe_dump(deployment.model_dump(), f)
+    save_as_yaml(deployment, layout.deployment_snapshot_path)
 
 
 def _backup_snapshot(layout: Layout) -> None:
@@ -32,9 +30,7 @@ def _backup_snapshot(layout: Layout) -> None:
     step.
     """
     if layout.deployment_snapshot_path.exists():
-        logger.debug(
-            "Backup previous snapshot to: %s", layout.previous_deployment_snapshot_path
-        )
+        logger.debug("Backup snapshot to: %s", layout.previous_deployment_snapshot_path)
         layout.deployment_snapshot_path.rename(layout.previous_deployment_snapshot_path)
 
 
@@ -43,8 +39,8 @@ def load_snapshot(layout: Layout, from_scratch: bool = False) -> Deployment:
 
     Args:
         layout: The ``Layout`` representing the Deployment Area.
-        from_scratch: If True, this will return the default ``Deployment`` configuration
-            in order to work with an empty Deployment Area.
+        from_scratch: If True, this will return the default ``Deployment``
+            configuration in order to work with an empty Deployment Area.
     """
     if from_scratch:
         if not layout.deployment_root.exists():
